Move Legend to SocialLocationPreview component since it's the only place that needs it.
