+++ src/components/HomepageUpsell/HomepageUpsell.tsx
@@ -35,11 +35,7 @@ const HomepageUpsell: React.FC<{ showHomepageUpsell: boolean }> = ({
                 to="/"
               >
                 <MapWrapper>
-                  <Map
-                    onClick={undefined}
-                    hideLegend={true}
-                    showCounties={false}
-                  />
+                  <Map onClick={undefined} showCounties={false} />
                 </MapWrapper>
                 <div>
                   <LabelWithChevron text="Risk &amp; vaccine tracker" />
