+++ src/components/SocialLocationPreview/Legend.style.ts
