+++ src/components/SocialLocationPreview/Legend.tsx
